# Add daemon architecture for high-performance speech-to-text