--- a/scribe-daemon.desktop
+++ b/scribe-daemon.desktop
@@ -1,0 +1,12 @@
+[Desktop Entry]
+Name=Scribe Daemon - Speech to Text
+Comment=High-performance speech-to-text transcription with daemon architecture
+Exec=uv run scribe-gui-daemon
+Path=/home/changeme/path/to/scribe
+Icon=audio-input-microphone
+Terminal=false
+Type=Application
+Categories=AudioVideo;Audio;Office;
+Keywords=speech;transcription;whisper;audio;voice;daemon;fast;
+StartupNotify=true
+MimeType=