… DAEMON_ARCHITECTURE.md
… src/scribe/client.py
… src/scribe/daemon.py
… src/scribe/daemon_manager.py
… src/scribe/gui_daemon.py
… src/scribe/ipc.py
… test_daemon.py